refactor: `domain` 패키지에 있던 user 도메인 관련 클래스를 `services/user/domain` 패키지 하위로 옮김
