+++ app/src/main/java/com/codesoom/assignment/infra/JpaUserRepository.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment.infra;
 
-import com.codesoom.assignment.domain.User;
+import com.codesoom.assignment.services.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
