+++ app/src/main/java/com/codesoom/assignment/services/user/UserRegisterService.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment.services.user;
 
-import com.codesoom.assignment.domain.User;
+import com.codesoom.assignment.services.user.domain.User;
 import com.codesoom.assignment.infra.JpaUserRepository;
 import org.springframework.stereotype.Service;
